Fix: The last configuration wasn't properly adjusted.
Changed the structure of this feature by deleting the
UserEventLike and UserEventLikeId entity and creating
a Like entity that has relationship with the Event and
User entities.

This commit introduces the creation of a Like entity,
which will enable users to engage with events by
liking them. This feature will improve user
experience and encourage participation in events.